requirements.txt
paddlepaddle-gpu==2.0.1
opencv-python==3.4.2.17
scikit-image==0.14.2
scipy==1.1.0